Crate kensetsu

Kensetsu is an over collateralized lending protocol, clone of MakerDAO. An individual can create a collateral debt positions (CDPs) for one of the listed token and deposit or lock amount of the token in CDP as collateral. Then the individual is allowed to borrow new minted Kensetsu USD (KUSD) in amount up to value of collateral corrected by liquidation_ratio coefficient. The debt in KUSD is a subject of stability_fee interest rate. Collateral may be unlocked only when the debt and the interest are paid back. If the value of collateral has changed in a way that it does not secure the debt, the collateral is liquidated to cover the debt and the interest.

Re-exports

pub use pallet::*;

Modules

migrations

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for kensetsu

Structs

CollateralInfo

Collateral parameters, includes risk info and additional data for interest rate calculation

CollateralRiskParameters

Risk management parameters for the specific collateral type.

CollateralizedDebtPosition

CDP - Collateralized Debt Position. It is a single collateral/debt record.

Enums

CdpType

Constants

TECH_ACCOUNT_PREFIX

TECH_ACCOUNT_TREASURY_MAIN