Module kensetsu::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Structs

DefaultLiquidatedThisBlock

Default value for LiquidatedThisBlock storage

Pallet

The pallet implementing the on-chain logic.

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configuration trait of this pallet.

Functions

__type_value_for_default_liquidated_this_block

Default value for LiquidatedThisBlock storage

Type Definitions

BadDebt

System bad debt, the amount of KUSD not secured with collateral.

BorrowTax

Risk parameter Borrows tax to buy back and burn KEN

CDPDepository

Storage of all CDPs, where key is a unique CDP identifier

CdpId

CDP id type

CdpOwnerIndex

Index links owner to CDP ids, not needed by protocol, but used by front-end

CollateralInfos

Parametes for collaterals, include risk parameters and interest recalculation coefficients

KusdHardCap

Risk parameter Hard cap of KUSD may be minted by the system

LiquidatedThisBlock

Flag indicates that liquidation took place in this block. Only one liquidation per block is allowed, the flag is dropped every block.

LiquidationPenalty

Liquidation penalty

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

NextCDPId

CDP counter used for CDP id

Timestamp